requirements.txt
alembic==1.16.5
blinker==1.9.0
click==8.2.1
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
Pillow 
Flask-JWT-Extended==4.7.1